+++ sample/reduxApp.js
@@ -1,0 +1,29 @@
+import {createStore} from 'redux';
+import * as Sentry from '@sentry/react';
+
+const initialState = {
+  counter: 0,
+};
+
+const reducer = (state = initialState, action) => {
+  switch (action.type) {
+    case 'COUNTER_INCREMENT':
+      return {
+        ...state,
+        counter: state.counter + 1,
+      };
+    case 'COUNTER_RESET':
+      return {
+        ...state,
+        counter: 0,
+      };
+    default:
+      return state;
+  }
+};
+
+const sentryEnhancer = Sentry.createReduxEnhancer();
+
+const store = createStore(reducer, sentryEnhancer);
+
+export {store};