+++ src/js/backend.ts
@@ -1,4 +1,4 @@
-import { BrowserOptions, Transports } from "@sentry/browser";
+import { BrowserOptions, Transports } from "@sentry/react";
 import { BrowserBackend } from "@sentry/browser/dist/backend";
 import { BaseBackend, NoopTransport } from "@sentry/core";
 import { Event, EventHint, Severity, Transport } from "@sentry/types";
@@ -97,7 +97,7 @@ export class ReactNativeBackend extends BaseBackend<BrowserOptions> {
 
     const transportOptions = {
       ...this._options.transportOptions,
-      dsn: this._options.dsn
+      dsn: this._options.dsn,
     };
 
     if (this._options.transport) {