+++ src/js/sdk.ts
@@ -1,4 +1,4 @@
-import { defaultIntegrations, getCurrentHub } from "@sentry/browser";
+import { defaultIntegrations, getCurrentHub } from "@sentry/react";
 import { initAndBind, setExtra } from "@sentry/core";
 import { Hub, makeMain } from "@sentry/hub";
 import { RewriteFrames } from "@sentry/integrations";
@@ -11,7 +11,7 @@ import {
   DebugSymbolicator,
   DeviceContext,
   ReactNativeErrorHandlers,
-  Release
+  Release,
 } from "./integrations";
 import { ReactNativeScope } from "./scope";
 
@@ -19,7 +19,7 @@ import { ReactNativeScope } from "./scope";
 
 const IGNORED_DEFAULT_INTEGRATIONS = [
   "GlobalHandlers", // We will use the react-native internal handlers
-  "TryCatch" // We don't need this
+  "TryCatch", // We don't need this
 ];
 
 /**
@@ -28,7 +28,7 @@ const IGNORED_DEFAULT_INTEGRATIONS = [
 export function init(
   options: ReactNativeOptions = {
     enableNative: true,
-    enableNativeCrashHandling: true
+    enableNativeCrashHandling: true,
   }
 ): void {
   const reactNativeHub = new Hub(undefined, new ReactNativeScope());
@@ -41,7 +41,7 @@ export function init(
       new Release(),
       ...defaultIntegrations.filter(
         (i) => !IGNORED_DEFAULT_INTEGRATIONS.includes(i.name)
-      )
+      ),
     ];
     if (__DEV__) {
       options.defaultIntegrations.push(new DebugSymbolicator());
@@ -55,7 +55,10 @@ export function init(
               .replace(/^address at /, "")
               .replace(/^.*\/[^\.]+(\.app|CodePush|.*(?=\/))/, "");
 
-            if (frame.filename !== "[native code]" && frame.filename !== "native") {
+            if (
+              frame.filename !== "[native code]" &&
+              frame.filename !== "native"
+            ) {
               const appPrefix = "app://";
               // We always want to have a triple slash
               frame.filename =