+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - fix: Use `LogBox` instead of `YellowBox` if possible. #989
 - fix: Don't add `DeviceContext` default integration if `enableNative` is set to `false`. #993
 - fix: Don't log "Native Sentry SDK is disabled" if `enableNativeNagger` is set to `false`. #993
+- feat: Migrate to `@sentry/react` from `@sentry/browser` and expose `ErrorBoundary` & the redux enhancer. #1005
 
 ## 1.6.3